Item Refactoring - Ensure Save on Sell/Delete

+++ Zolian.Server.Base/Types/Inventory.cs
@@ -101,12 +101,22 @@ public class Inventory : ObjectManager, IInventory
     public void RemoveFromInventory(WorldClient client, Item item)
     {
         if (item == null) return;
-
-        if (Items.TryUpdate(item.InventorySlot, null, item))
-            client.SendRemoveItemFromPane(item.InventorySlot);
-        client.LastItemDropped = item;
-        UpdatePlayersWeight(client);
-        item.DeleteFromAislingDb();
+        
+        // Remove item from cache prior to delete
+        if(ServerSetup.Instance.GlobalSqlItemCache.TryRemove(item.ItemId, out _))
+        {
+            if (Items.TryUpdate(item.InventorySlot, null, item))
+                client.SendRemoveItemFromPane(item.InventorySlot);
+            client.LastItemDropped = item;
+            UpdatePlayersWeight(client);
+            item.DeleteFromAislingDb();
+            // Ensure item was removed from cache, routine can sometimes add it back
+            ServerSetup.Instance.GlobalSqlItemCache.TryRemove(item.ItemId, out _);
+        }
+        else
+        {
+            ServerSetup.Logger($"Issue removing item {item.ItemId} from {client.RemoteIp} - {client.Aisling.Serial}");
+        }
     }
 
     /// <summary>
